backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.1
python-multipart>=0.0.9
pillow>=10.4.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
pydantic[email]>=2.6.4
python-dotenv>=1.0.1
httpx>=0.27.0
opencv-python-headless>=4.9.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
